fix: allow direct Node launches of GPT-Live examples
