+++ examples/src/gpt_live/client_delegation.ts
@@ -13,7 +13,29 @@ import {
 } from '@livekit/agents';
 import * as openai from '@livekit/agents-plugin-openai';
 import { fileURLToPath } from 'node:url';
-import { checkOrderStatus, lookupWeather } from './tools.js';
+import { z } from 'zod';
+
+const orders: Record<string, string> = {
+  A1042: 'shipped, arriving Thursday',
+  B2231: 'still being packed',
+};
+
+const checkOrderStatus = llm.tool({
+  name: 'checkOrderStatus',
+  description: 'Check the delivery status of an order.',
+  parameters: z.object({ orderId: z.string().describe('The order reference, such as A1042.') }),
+  execute: async ({ orderId }) => {
+    const status = orders[orderId.toUpperCase()];
+    return status ? `Order ${orderId} is ${status}.` : `There is no order ${orderId}.`;
+  },
+});
+
+const lookupWeather = llm.tool({
+  name: 'lookupWeather',
+  description: 'Look up the current weather for a location.',
+  parameters: z.object({ location: z.string().describe('The city or region to look up.') }),
+  execute: async ({ location }) => `The weather in ${location} is 62 degrees and partly cloudy.`,
+});
 
 const deskTools = new llm.ToolContext([checkOrderStatus, lookupWeather]);
 const deskInstructions =
